WeatherResponseをWeatherForecastResponseに変更。
